refactor of database to use bigint (long) for timeStamps. all fields for storage of timeStamp has been renamed to timeStamp and startTimeStamp/endTimeStamp Methods in the repository has been changed to reflect these changes. Mock objects has been updated to reflect this change as well.
git-svn-id: http://subversion.assembla.com/svn/drunkendroid/src@212 3995579f-fdfd-45ad-a057-9d0aee8b8447

+++ drunkendroid-server/src/itu/malta/drunkendroidserver/util/xstreem/converters/MoodMapConverter.java
@@ -21,7 +21,7 @@ public class MoodMapConverter implements Converter{
 	public void marshal(Object value, HierarchicalStreamWriter writer,
 			MarshallingContext context) {
 		  GridCell MMReading = (GridCell) value;
-		  writer.addAttribute("value", Integer.toString(MMReading.getAverage()));
+		  writer.addAttribute("value", Long.toString(MMReading.getAverage()));
 		  writer.addAttribute("long", Double.toString(MMReading.getLongitude()));
 		  writer.addAttribute("lat", Double.toString(MMReading.getLatitude()));
 	}
